refactor: Use common folder structure
- Source files are under src/
- Tests are part of the src/ folder
  - Makes the spec file easier to find
  - Easier to see which files are lacking test
  - No need to deal with complex relative paths in tests
- Did not retain the lib folder, tried to flatten the directory
  structure
- Utils folder contains code that is not particularly relevant to the
  project
