--- packages/cozy-app-publish/index.js
@@ -1,7 +1,0 @@
-const travisScript = require('./lib/travis')
-const manualScript = require('./lib/manual')
-
-module.exports = {
-  travis: travisScript,
-  manual: manualScript
-}
