+++ packages/cozy-app-publish/package.json
@@ -1,7 +1,7 @@
 {
   "name": "cozy-app-publish",
   "version": "0.22.1",
-  "main": "index.js",
+  "main": "src/index.js",
   "author": "Cozy",
   "license": "MIT",
   "repository": {
@@ -17,7 +17,7 @@
     "test:integration": "env TEST_INTEGRATION=1 jest"
   },
   "bin": {
-    "cozy-app-publish": "./cozy-app-publish.js"
+    "cozy-app-publish": "./src/cozy-app-publish.js"
   },
   "dependencies": {
     "argparse": "^1.0.10",
@@ -45,7 +45,7 @@
   },
   "jest": {
     "setupFiles": [
-      "./test/jestLibs/setupJest.js"
+      "./src/test/jest.setup.js"
     ]
   },
   "engines": {
