+++ packages/cozy-app-publish/src/__mocks__/errored-hook.js
