+++ packages/cozy-app-publish/src/__mocks__/hook.js
