+++ packages/cozy-app-publish/src/__mocks__/prepublish-bad-value-options-hook.js
