+++ packages/cozy-app-publish/src/__mocks__/prepublish-missing-manifest-options-hook.js
