+++ packages/cozy-app-publish/src/__mocks__/prepublish-missing-options-hook.js
