+++ packages/cozy-app-publish/src/__mocks__/prepublish-unsanitized-hook.js
