+++ packages/cozy-app-publish/src/constants.js
