+++ packages/cozy-app-publish/src/cozy-app-publish.js
