+++ packages/cozy-app-publish/src/hooks/pre/__mocks__/helpers.js
