+++ packages/cozy-app-publish/src/hooks/pre/downcloud.js
@@ -8,7 +8,7 @@ const {
   createArchive,
   pushArchive
 } = require('./helpers')
-const logger = require('../../../utils/logger')
+const logger = require('../../utils/logger')
 
 module.exports = async options => {
   if (!fs.existsSync(options.buildDir)) {
