+++ packages/cozy-app-publish/src/hooks/pre/helpers.js
