+++ packages/cozy-app-publish/src/hooks/pre/helpers.spec.js
