+++ packages/cozy-app-publish/src/index.js
@@ -1,0 +1,7 @@
+const travisScript = require('./travis')
+const manualScript = require('./manual')
+
+module.exports = {
+  travis: travisScript,
+  manual: manualScript
+}
