+++ packages/cozy-app-publish/src/postpublish.js
@@ -1,4 +1,4 @@
-const runHooks = require('../utils/runhooks')
+const runHooks = require('./runhooks')
 
 module.exports = async options =>
   runHooks(options.postpublishHook, 'post', options)
