+++ packages/cozy-app-publish/src/publish.js
