+++ packages/cozy-app-publish/src/publisher.js
@@ -1,11 +1,11 @@
 const postpublish = require('./postpublish')
 const prepublish = require('./prepublish')
 const publish = require('./publish')
-const colorize = require('../utils/colorize')
+const colorize = require('./utils/colorize')
 const constants = require('./constants')
 const promptConfirm = require('./confirm')
 const defaults = require('lodash/defaults')
-const logger = require('../utils/logger')
+const logger = require('./utils/logger')
 
 const {
   DEFAULT_REGISTRY_URL,
