+++ packages/cozy-app-publish/src/runhooks.js
@@ -1,7 +1,7 @@
-const logger = require('./logger')
+const logger = require('./utils/logger')
 
 const requireHook = (hook, type) => {
-  const builtInHookPath = `${__dirname}/../lib/hooks/${type}/` + hook + '.js'
+  const builtInHookPath = `${__dirname}/hooks/${type}/` + hook + '.js'
   const customHookPath = process.cwd() + '/' + hook
   try {
     return require(builtInHookPath)
