+++ packages/cozy-app-publish/src/tags.js
