+++ packages/cozy-app-publish/src/tags.spec.js
