+++ packages/cozy-app-publish/src/test/jest.setup.js
@@ -3,7 +3,7 @@
 const fetch = require('jest-fetch-mock')
 jest.doMock('node-fetch', () => fetch.mockResponse({ status: 201 }))
 
-jest.mock('../../utils/logger', () => ({
+jest.mock('../utils/logger', () => ({
   log: jest.fn(),
   warn: jest.fn(),
   error: jest.fn()
