+++ packages/cozy-app-publish/src/travis.js
@@ -1,6 +1,6 @@
 const path = require('path')
-const getManifestAsObject = require('../utils/getManifestAsObject')
-const getTravisVariables = require('../utils/getTravisVariables')
+const getManifestAsObject = require('./utils/getManifestAsObject')
+const getTravisVariables = require('./utils/getTravisVariables')
 const tags = require('./tags')
 const { getDevVersion } = require('./tags')
 const publisher = require('./publisher')
