+++ packages/cozy-app-publish/src/utils/colorize.js
