+++ packages/cozy-app-publish/src/utils/getManifestAsObject.js
