+++ packages/cozy-app-publish/src/utils/getTravisVariables.js
